+++ bqat/app.py
@@ -29,6 +29,7 @@ from bqat.utils import (
     fix_filepath,
     generate_report,
     iter_matching_files,
+    reconstruct_filepath,
     split_input_folder,
     validate_path,
     write_csv,
@@ -54,6 +55,7 @@ async def run(
     query: str,
     sort: str,
     cwd: str,
+    prefix: str,
     batch: int,
     fusion: int,
     engine: str,
@@ -97,7 +99,7 @@ async def run(
         1 for _ in iter_matching_files(input_folder, pattern, extended(TYPE))
     )
 
-    metadata.add_row("Input Folder", f"[dark_goldenrod]{str(input_folder)}")
+    metadata.add_row("Input Folder", f"[dark_goldenrod]{prefix + input_folder}")
     metadata.add_row("Input Count", f"[bold yellow]{str(file_total)}")
 
     console.print(
@@ -170,7 +172,8 @@ async def run(
                         convert,
                         target,
                         engine,
-                        fusion,
+                        fusion=fusion,
+                        prefix=prefix,
                     )
                 )
                 not_ready = True
@@ -205,6 +208,7 @@ async def run(
                         convert,
                         target,
                         engine,
+                        prefix=prefix,
                     )
                 )
                 file_count += 1
@@ -286,7 +290,7 @@ async def run(
     try:
         if output_dir and reporting:
             # write_report(report_dir, output_dir, f"EDA Report (BQAT v{version})")
-            dir = generate_report(output_dir, cwd)
+            dir = generate_report(output_dir, cwd, prefix)
             report_dir = (
                 {"Preview Table": dir.get("table"), "EDA Report": dir.get("report")}
                 if dir
@@ -300,7 +304,7 @@ async def run(
 
     try:
         if output_dir and (attributes or query or sort):
-            dir = filter_output(output_dir, attributes, query, sort, cwd)
+            dir = filter_output(output_dir, attributes, query, sort, cwd, prefix)
             outlier_filter = (
                 {"Output": dir.get("output"), "Report": dir.get("report")}
                 if dir
@@ -332,9 +336,9 @@ async def run(
     Console().print("\n>> [bright_yellow]Task Finished[/bright_yellow] <<\n")
 
 
-def filter(output, attributes, query, sort, cwd):
+def filter(output, attributes, query, sort, cwd, prefix):
     try:
-        dir = filter_output(output, attributes, query, sort, cwd)
+        dir = filter_output(output, attributes, query, sort, cwd, prefix)
         outlier_filter = (
             {
                 "Table": dir.get("table"),
@@ -352,7 +356,7 @@ def filter(output, attributes, query, sort, cwd):
         print("\n> Summary:")
         summary = {"Output Filter": outlier_filter}
         Console().print_json(json.dumps(summary))
-    print("\n>> [bright_yellow]Task Finished[/bright_yellow] <<\n")
+    Console().print("\n>> [bright_yellow]Task Finished[/bright_yellow] <<\n")
     return dir
 
 
@@ -542,11 +546,22 @@ async def benchmark(
     Console().print_json(json.dumps(summary))
     # with open("data/benchmark.json", "w") as f:
     #     json.dump(result, f)
-    print("\n>> Benchmarking Finished <<\n")
+    Console().print("\n>> Benchmarking Finished <<\n")
 
 
 @ray.remote
-def scan_task(path, output_dir, log_dir, mode, convert, target, engine, fusion=6):
+def scan_task(
+    path,
+    output_dir,
+    log_dir,
+    mode,
+    convert,
+    target,
+    engine,
+    fusion=6,
+    prefix="",
+):
+    results = []
     if mode == "speech" or (mode == "face" and engine in ("ofiq", "fusion")):
         try:
             result = scan(path, mode=mode, engine=engine, fusion=fusion)
@@ -557,11 +572,13 @@ def scan_task(path, output_dir, log_dir, mode, convert, target, engine, fusion=6
             return
 
         for result in result_list:
-            write_csv(output_dir, fix_filepath(result))
+            result = fix_filepath(result)
+            result = reconstruct_filepath(result, prefix)
             if result.get("log"):
                 log_dict = {"folder": path, "logs": result.pop("log")}
                 write_log(log_dir, log_dict)
-        return result_list
+            write_csv(output_dir, result)
+            results.append(result)
     else:
         try:
             result = scan(path, mode=mode, source=convert, target=target, engine=engine)
@@ -575,9 +592,10 @@ def scan_task(path, output_dir, log_dir, mode, convert, target, engine, fusion=6
             for log in logs:
                 log.update({"file": path})
                 write_log(log_dir, log)
-
+        result = reconstruct_filepath(result, prefix)
         write_csv(output_dir, result)
-        return [result]
+        results.append(result)
+    return results
 
 
 @ray.remote
@@ -598,9 +616,9 @@ def benchmark_task(path: str, mode: str, engine: str, fusion=6) -> None:
         )
 
 
-def report(input, cwd):
+def report(input, cwd, prefix):
     try:
-        dir = generate_report(input, cwd)
+        dir = generate_report(input, cwd, prefix)
         report = (
             {"Preview Table": dir.get("table"), "EDA Report": dir.get("report")}
             if dir
@@ -614,7 +632,7 @@ def report(input, cwd):
         print("\n> Summary:")
         summary = {"EDA Report": report}
         Console().print_json(json.dumps(summary))
-    print("\n>> [bright_yellow]Task Finished[/bright_yellow] <<\n")
+    Console().print("\n>> [bright_yellow]Task Finished[/bright_yellow] <<\n")
     return dir
 
 
@@ -753,12 +771,15 @@ def preprocess(input_dir: str, output_dir: str, debugging: bool, config: dict) -
         },
     }
     Console().print_json(json.dumps(summary))
-    print("\n>> [bright_yellow]Preprocessing Task Finished[/bright_yellow] <<\n")
+    Console().print(
+        "\n>> [bright_yellow]Preprocessing Task Finished[/bright_yellow] <<\n"
+    )
 
 
 @ray.remote
 def preprocess_task(file: str, output: dir, config: dict) -> None:
     try:
+        import wsq
         file = Path(file)
         if not Path(output).exists():
             Path(output).mkdir(parents=True, exist_ok=True)